refactor: move from core util with ifdef to platform implementations
