+++ platform/android/android.cmake
@@ -70,6 +70,7 @@ target_sources(
         ${PROJECT_SOURCE_DIR}/platform/default/src/mln/text/bidi.cpp
         ${PROJECT_SOURCE_DIR}/platform/default/src/mln/util/compression.cpp
         ${PROJECT_SOURCE_DIR}/platform/default/src/mln/util/filesystem.cpp
+        ${PROJECT_SOURCE_DIR}/platform/default/src/mln/util/i18n.cpp
         ${PROJECT_SOURCE_DIR}/platform/default/src/mln/util/monotonic_timer.cpp
         ${PROJECT_SOURCE_DIR}/platform/default/src/mln/util/png_writer.cpp
         ${PROJECT_SOURCE_DIR}/platform/default/src/mln/util/thread_local.cpp
